moved edteput interpreter ro its class
